+++ src/parser.nim
@@ -19,11 +19,12 @@ type
     ntGrouping,
     ntArray,
     ntQuotation,
-    ntBinding
+    ntBinding,
+    ntDebug
 
   Node* = ref object
     case typ: NodeType
-    of ntIdent, ntOperator, ntReal, ntString:
+    of ntIdent, ntOperator, ntReal, ntString, ntDebug:
       tok: Token
     of ntGrouping, ntArray:
       anchor: Token
@@ -62,7 +63,7 @@ func name*(self: Node): ptr Token =
 
 func trace*(self: Node): string =
   case self.typ
-  of ntIdent, ntOperator, ntReal, ntString:
+  of ntIdent, ntOperator, ntReal, ntString, ntDebug:
     self.tok.trace()
   of ntGrouping, ntArray:
     self.anchor.trace()
@@ -73,7 +74,7 @@ func trace*(self: Node): string =
 
 func lit*(self: Node): string =
   case self.typ
-  of ntIdent, ntOperator, ntReal, ntString: 
+  of ntIdent, ntOperator, ntReal, ntString, ntDebug: 
      self.tok.lit
   of ntGrouping, ntArray:
     "(" & self.nodes.mapIt(it.lit).join(" ") & ")"
@@ -87,7 +88,7 @@ proc `$`*(self: Node): string =
   result = fmt"({self.typ} "
 
   case self.typ
-  of ntIdent, ntOperator, ntReal, ntString: 
+  of ntIdent, ntOperator, ntReal, ntString, ntDebug: 
     result &= self.tok.min()
   of ntGrouping, ntArray:
     result &= self.nodes.mapIt($it).join(" ")
@@ -137,6 +138,8 @@ proc parseItem(self: Parser, topLevel: bool = false): Node =
     result = self.parseGrouping(tok)
   of ttBracketLeft:
     result = self.parseArray(tok)
+  of ttDebug:
+    result = Node(typ: ntDebug, tok: tok[])
   else:
     let e = newVernError(fmt"Unexpected token '{tok[].lit}'")
     e.addTrace(tok[].trace())